Address Copilot review round 3 on #537
- evm/rpc: redact the endpoint in the parity warning — provider RPC URLs often
  embed API keys in the path/query, so log a redacted form (Copilot security).
  Reuse the canonical redactUrl, extracted to a neutral `src/redact.ts` so
  evm/rpc need not depend on the fallback module; fallback/diagnostics re-exports
  it to keep the public surface stable.
- test: fix a stale comment referencing require('../rpc') (now '../rpc/builder').
